Local Fisher Discriminant Analysis of Kernel principle components (LFDAKPC)

Local Fisher Discriminant Analysis of Kernel principle components

LFDAKPC(x, y, n.pc, usekernel = FALSE, fL = 0, kernel.name = "rbfdot", kpar = list(0.001), kernel = "gaussian", threshold = 1e-05, ...)

Arguments

x
Input traing data

y
Input labels

n.pc
number of pcs that will be kept in analysis

usekernel
Whether to use kernel function, if TRUE, it will pass to the kernel.names

fL
if using kernel, pass to kernel function

kernel.name
if usekernel is TURE, this will take the kernel name and use the parameters set as you defined

kpar
the list of hyper-parameters (kernel parameters). This is a list which contains the parameters to be used with the kernel function. Valid parameters for existing kernels are :

sigma inverse kernel width for the Radial Basis kernel function "rbfdot" and the Laplacian kernel "laplacedot".

degree, scale, offset for the Polynomial kernel "polydot"

scale, offset for the Hyperbolic tangent kernel function "tanhdot"

sigma, order, degree for the Bessel kernel "besseldot".

sigma, degree for the ANOVA kernel "anovadot".

Hyper-parameters for user defined kernels can be passed through the kpar parameter as well.

kernel
kernel name if all the above are not used

threshold
the threshold for kpc: value of the eigenvalue under which principal components are ignored (only valid when features = 0). (default : 0.0001)

Details

Sugiyama, M (2007). Dimensionality reduction of multimodal labeled data by local Fisher discriminant analysis. Journal of Machine Learning Research, vol.8, 1027-1061.

Sugiyama, M (2006). Local Fisher discriminant analysis for supervised dimensionality reduction. In W. W. Cohen and A. Moore (Eds.), Proceedings of 23rd International Conference on Machine Learning (ICML2006), 905-912.

Tang, Y., & Li, W. (2019). lfda: Local Fisher Discriminant Analysis inR. Journal of Open Source Software, 4(39), 1572.

Karatzoglou, A., Smola, A., Hornik, K., & Zeileis, A. (2004). kernlab-an S4 package for kernel methods in R. Journal of statistical software, 11(9), 1-20.

Value

kpca

Results of kernel principal component analysis. Kernel Principal Components Analysis is a nonlinear form of principal component analysis

kpc

Kernel principal components. The scores of the components

LFDAKPC

LOcal linear discriminant anslysis of kernel principal components

LDs

The discriminant function. The scores of the components

label

The corresponding class of the data

n.pc

Number of Pcs kept in analysis

References

Author

qinxinghu@gmail.com

Examples

train=LFDAKPC(iris[,1:4],y=iris[,5],tol=1,n.pc=3,kernel.name = "rbfdot")
pred=predict.LFDAKPC(train,prior=NULL,testData = iris[1:10,1:4])