Prediction of Local Fisher Discriminant Analysis of Kernel principle components (LFDAKPC)

This is the prediction function of Local Fisher Discriminant Analysis of Kernel principle components (LFDAKPC)

predict.LFDAKPC(object = obj, prior, testData = data)

Arguments

object
LFDAKPC object

prior
The prior proportion or weight

testData
The test data

Details

Value

predicted_LDs

predicted discriminant function

predict.LFDAKPC

predicted results of the test data

%% ...
References

Karatzoglou, A., Smola, A., Hornik, K., & Zeileis, A. (2004). kernlab-an S4 package for kernel methods in R. Journal of statistical software, 11(9), 1-20.

Sugiyama, M (2007). Dimensionality reduction of multimodal labeled data by local Fisher discriminant analysis. Journal of Machine Learning Research, vol.8, 1027-1061.

Sugiyama, M (2006). Local Fisher discriminant analysis for supervised dimensionality reduction. In W. W. Cohen and A. Moore (Eds.), Proceedings of 23rd International Conference on Machine Learning (ICML2006), 905-912.

Tang, Y., & Li, W. (2019). lfda: Local Fisher Discriminant Analysis inR. Journal of Open Source Software, 4(39), 1572.

Karatzoglou, A., Smola, A., Hornik, K., & Zeileis, A. (2004). kernlab-an S4 package for kernel methods in R. Journal of statistical software, 11(9), 1-20.

Author

qinxinghu@gmail.com

Examples

train=LFDAKPC(iris[,1:4],y=iris[,5],tol=1,n.pc=3,kernel.name = "rbfdot")
pred=predict.LFDAKPC(train,prior=NULL,testData = iris[1:10,1:4])